Refactor all math functions to MathUtils.js
